Allow nullable fields
The `field` method is constrained to `Any`. However, in some cases, it's possible that one wants to construct a null value depending on a certain logic, for example, a LocalDate represented by the String "00000000" should be null as per `field(0, 8) { if(equals("00000000")) null else toMyLocalDate()  }`

This commit changes that behavior to enable parsing nullable types using `field`.

The `parseToType`, `defaultTypeParser` methods were changed in a way that their tests had to change as well. This won't break current users as these functions are internal and not intended for external users.

Closes #8