Add LineParseException (#18)
Sometimes, when parsing a line with some sort of unexpected content, is hard to track down the line where this occurred, specially on large files.

This pull request adds the LineParseException, which encapsulates all sort of errors which might occur when trying to parse any line. If necessary, it will be possible to access the line which generated the error.